refactor(resume-core): split calculations.js into focused modules (behavior-preserving)
